VMware: Move helper functions from vmware_host_config_manager to module util.
